pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "pansh"
dynamic = ["version"]
description = "AnyShare CLI for browsing, searching, uploading, and downloading files"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "LZR" }]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Environment :: Console",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "httpx>=0.25,<1",
  "platformdirs>=3.0,<5",
  "prompt-toolkit>=3.0.36,<4",
  "pydantic>=2.4,<3",
  "pyyaml>=6,<7",
  "rich>=13.5,<14",
  "rsa>=4.9,<5",
  "typer>=0.9,<1",
]

[project.optional-dependencies]
dev = [
  "build>=1.0",
  "mypy>=1.8",
  "pytest>=7.4",
  "ruff>=0.3",
  "twine>=5",
]

[project.scripts]
pansh = "pansh.main:main"
bhpan = "pansh.main:main"

[tool.hatch.version]
path = "src/pansh/version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/pansh"]

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
  "/README.md",
  "/pyproject.toml",
  "/LICENSE",
]

[tool.pytest.ini_options]
norecursedirs = [
  ".venv",
  ".pytest_cache",
  ".pytest-tmp",
  ".tmp",
  ".tmp-settings",
  "dist",
  "build",
  "pytest-cache-files-*",
]
